Finished Accounts (mostly) + refinements
Accounts now have required parameters and optional parameters (allowing some null values to be saved). The current user's username is saved to the account they created. Added search function from the Journal but it does not currently work. Journal entries also do not work for me. Also: All migrations have been archived to the "Archive" folder in the root. There is now one Initial migration because I was having issues with the others.